move INotification into arcgis-rest-users
